% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGE_DOCUMENTATION.R
\docType{package}
\name{cbpdata}
\alias{cbpdata}
\alias{cbpdata-package}
\title{cbpdata: Congresional Bills Bulk Data Collection}
\description{
cbpdata: Congresional Bills Bulk Data Collection
}
\section{cbpdata functions}{

This package provides functions to generate the congressional bills data,
and to update those data from source files on govinfo.gov. This help file
will be fleshed out over time.
}

