% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill_list.R
\name{get_bill_list}
\alias{get_bill_list}
\title{Generate list of Bills to Collect}
\usage{
get_bill_list(start_year = NULL, end_year = NULL)
}
\arguments{
\item{start_year}{Defaults to NULL, in which case results are returned for
the current calendar year. Can be any number greater than or equal to 1992,
in which case the results for all years in that range will be returned. If
not NULL, must be of the form `1994` or `2019`}

\item{end_year}{Defaults to NULL, in which case results are returned for
the current calendar year. Can be any number greater than or equal to 1992,
in which case the results for all years in that range will be returned. If
not NULL, must be of the form `1994` or `2019`}
}
\value{
A list object with `bill_metadata`, `bill_text`, and `year` fields
where the first two fields contains a vector of urls and the `year` field
contains a vector of years that correspond to the returned URLS. Also
contains a `last_modified` field with a vector of dates the relevant pages
were last modified.
}
\description{
Downloads sitemaps for specified year range and generates file
lists to be downloaded.
}
